Restrict Date type precision to system_clock precision in Swift
C++ type std::chrono::system_clock::time_point has microsecond precision
on iOS/MacOS (as opposed to nanosecond precision of the same type on
Linux/Android).

Moved the Double<->Int64 conversion logic from Swift to CBridge. This
way the conversion code has access to std::chrono::system_clock::period
and can adjust conversion precision accordingly.

Related: APIGEN-1610

+++ genium/src/main/java/com/here/genium/generator/cbridge/CppTypeInfo.kt
@@ -69,8 +69,8 @@ open class CppTypeInfo(
             )
         )
         val DATE = CppTypeInfo(name = "std::chrono::system_clock::time_point",
-            cType = CType.INT64,
-            functionReturnType = CType.INT64,
+            cType = CType.DOUBLE,
+            functionReturnType = CType.DOUBLE,
             typeCategory = TypeCategory.BUILTIN_DATE,
             includes = listOf(
                 CppLibraryIncludes.CHRONO,
